requirements.txt
Flask
pandas
numpy
numpy-financial
openpyxl
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
PyJWT>=2.8.0
psycopg2-binary
python-dotenv